# frontend/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "VinylImportRequest": {
        "properties": {
          "acquired_on": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Acquired On"
          },
          "force": {
            "default": false,
            "title": "Force",
            "type": "boolean"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "release_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Release Id"
          },
          "sides": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/VinylSideInput"
            },
            "title": "Sides",
            "type": "array"
          },
          "url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Url"
          }
        },
        "title": "VinylImportRequest",
        "type": "object"
      },
      "VinylImportResponse": {
        "properties": {
          "bpms_applied": {
            "title": "Bpms Applied",
            "type": "integer"
          },
          "release": {
            "$ref": "#/components/schemas/VinylReleaseSummary"
          },
          "sides_written": {
            "title": "Sides Written",
            "type": "integer"
          },
          "unplannable": {
            "title": "Unplannable",
            "type": "integer"
          }
        },
        "required": [
          "release",
          "sides_written",
          "bpms_applied",
          "unplannable"
        ],
        "title": "VinylImportResponse",
        "type": "object"
      },
      "VinylPreviewResponse": {
        "properties": {
          "album": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Album"
          },
          "already_owned": {
            "default": false,
            "title": "Already Owned",
            "type": "boolean"
          },
          "artist": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Artist"
          },
          "catalog_number": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Catalog Number"
          },
          "country": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Country"
          },
          "cover_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cover Url"
          },
          "format": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Format"
          },
          "genre": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Genre"
          },
          "is_pressing": {
            "default": true,
            "title": "Is Pressing",
            "type": "boolean"
          },
          "label": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Label"
          },
          "rows": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/VinylPreviewRow"
            },
            "title": "Rows",
            "type": "array"
          },
          "source": {
            "title": "Source",
            "type": "string"
          },
          "source_id": {
            "title": "Source Id",
            "type": "string"
          },
          "year": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Year"
          }
        },
        "required": [
          "source",
          "source_id"
        ],
        "title": "VinylPreviewResponse",
        "type": "object"
      },
...
      "VinylReleaseDetail": {
        "properties": {
          "release": {
            "$ref": "#/components/schemas/VinylReleaseSummary"
          },
          "sides": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/VinylSide"
            },
            "title": "Sides",
            "type": "array"
          }
        },
        "required": [
          "release"
        ],
        "title": "VinylReleaseDetail",
        "type": "object"
      },
...
      "VinylSearchResponse": {
        "properties": {
          "kind": {
            "title": "Kind",
            "type": "string"
          },
          "query": {
            "title": "Query",
            "type": "string"
          },
          "results": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/VinylSearchResult"
            },
            "title": "Results",
            "type": "array"
          }
        },
        "required": [
          "query",
          "kind"
        ],
        "title": "VinylSearchResponse",
        "type": "object"
      },
...
      "VinylSidePatch": {
        "description": "A number the DJ typed. Always lands as `manual`, the top of the ladder.",
        "properties": {
          "bpm": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bpm"
          },
          "key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Key"
          },
          "length": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Length"
          }
        },
        "title": "VinylSidePatch",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Kiku API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/vinyl/import": {
      "post": {
        "description": "Write the pressing and its sides, with whatever numbers the DJ confirmed.",
        "operationId": "vinyl_import_api_vinyl_import_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VinylImportRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VinylImportResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Vinyl Import",
        "tags": [
          "vinyl"
        ]
      }
    },
    "/api/vinyl/preview": {
      "get": {
        "description": "The tracklist, with every BPM Kiku can work out before you type one.\n\n`url` accepts a Discogs or Bandcamp release link, for records the search\ncan't find \u2014 which is most white labels.",
        "operationId": "vinyl_preview_api_vinyl_preview_get",
        "parameters": [
          {
            "in": "query",
            "name": "release_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Release Id"
            }
          },
          {
            "in": "query",
            "name": "url",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Url"
            }
          },
          {
            "in": "query",
            "name": "fill_bpm",
            "required": false,
            "schema": {
              "default": true,
              "title": "Fill Bpm",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VinylPreviewResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Vinyl Preview",
        "tags": [
          "vinyl"
        ]
      }
    },
...
    "/api/vinyl/releases/{release_id}": {
      "delete": {
        "description": "Take a record off the shelf, with the sides that came in with it.",
        "operationId": "vinyl_release_delete_api_vinyl_releases__release_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "release_id",
            "required": true,
            "schema": {
              "title": "Release Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Vinyl Release Delete",
        "tags": [
          "vinyl"
        ]
      },
      "get": {
        "description": "One record, the sides on it in pressing order, and which you own as files.\n\nSuggestions are worked out on read and never stored: a guess that goes stale\nwhen the library changes is worse than one recomputed in a tenth of a second.",
        "operationId": "vinyl_release_detail_api_vinyl_releases__release_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "release_id",
            "required": true,
            "schema": {
              "title": "Release Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VinylReleaseDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Vinyl Release Detail",
        "tags": [
          "vinyl"
        ]
      }
    },
    "/api/vinyl/releases/{release_id}/digital": {
      "get": {
        "description": "The files behind a record, in pressing order \u2014 what \"play\" plays.",
        "operationId": "vinyl_release_digital_api_vinyl_releases__release_id__digital_get",
        "parameters": [
          {
            "in": "path",
            "name": "release_id",
            "required": true,
            "schema": {
              "title": "Release Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/TrackResponse"
                  },
                  "title": "Response Vinyl Release Digital Api Vinyl Releases  Release Id  Digital Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Vinyl Release Digital",
        "tags": [
          "vinyl"
        ]
      }
    },
    "/api/vinyl/search": {
      "get": {
        "description": "Find a pressing. One Discogs call, everything the picker needs.",
        "operationId": "vinyl_search_api_vinyl_search_get",
        "parameters": [
          {
            "description": "Catalogue number, barcode, artist, or title",
            "in": "query",
            "name": "q",
            "required": true,
            "schema": {
              "description": "Catalogue number, barcode, artist, or title",
              "minLength": 1,
              "title": "Q",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "vinyl_only",
            "required": false,
            "schema": {
              "default": true,
              "title": "Vinyl Only",
              "type": "boolean"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 12,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VinylSearchResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Vinyl Search",
        "tags": [
          "vinyl"
        ]
      }
    },
    "/api/vinyl/sides/{track_id}": {
      "patch": {
        "description": "Correct a side by hand.\n\nAnything set here is `manual` \u2014 it outranks every automatic source, so a\nwrong estimate stays corrected.",
        "operationId": "vinyl_side_patch_api_vinyl_sides__track_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "track_id",
            "required": true,
            "schema": {
              "title": "Track Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VinylSidePatch"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VinylSide"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Vinyl Side Patch",
        "tags": [
          "vinyl"
        ]
      }
    },
    "/api/vinyl/sides/{track_id}/twin": {
      "put": {
        "description": "This side is that file. Its BPM and key come from your analysis unless you typed them.",
        "operationId": "vinyl_side_link_api_vinyl_sides__track_id__twin_put",
        "parameters": [
          {
            "in": "path",
            "name": "track_id",
            "required": true,
            "schema": {
              "title": "Track Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VinylTwinRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VinylSide"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Vinyl Side Link",
        "tags": [
          "vinyl"
        ]
      }
    },
    "/api/vinyl/sides/{track_id}/twin/{digital_track_id}": {
      "delete": {
        "description": "\"Not it\" \u2014 unlinks the file if it was linked, and never suggests it again.",
        "operationId": "vinyl_side_unlink_api_vinyl_sides__track_id__twin__digital_track_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "track_id",
            "required": true,
            "schema": {
              "title": "Track Id",
              "type": "integer"
            }
          },
          {
            "in": "path",
            "name": "digital_track_id",
            "required": true,
            "schema": {
              "title": "Digital Track Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VinylSide"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Vinyl Side Unlink",
        "tags": [
          "vinyl"
        ]
      }
    },
...
  }
}